compilable for amd64, almost all the 'signed/unsigned mismatch' warnings suppressed
git-svn-id: http://merlin.fit.vutbr.cz/svn/STK/branches/64bit@70 1d61b05a-ac01-0410-9a19-bfe7c875843c

+++ src/STKLib/Models.h
@@ -304,7 +304,7 @@ namespace STK
     
     int                       mInputVectorSize;
     int                       mParamKind;
-    long                      mNMixtures;
+    size_t                    mNMixtures;
     size_t                    mNStates;
     bool                      mAllocAccums;
     int                       mTotalDelay;
